+++ src/main/java/jpa/jpa_shop/service/IFS/MemberServiceIFS.java
@@ -1,8 +1,8 @@
 package jpa.jpa_shop.service.IFS;
 
 import jpa.jpa_shop.domain.member.Member;
-import jpa.jpa_shop.web.controller.dto.request.member.MemberUpdateRequestDto;
-import jpa.jpa_shop.web.controller.dto.response.member.MemberResponseDto;
+import jpa.jpa_shop.web.dto.request.member.MemberUpdateRequestDto;
+import jpa.jpa_shop.web.dto.response.member.MemberResponseDto;
 
 import java.util.List;
