+++ src/main/java/jpa/jpa_shop/service/ItemService.java
@@ -8,7 +8,7 @@ import jpa.jpa_shop.domain.item.Repository.ItemRepository;
 import jpa.jpa_shop.exception.NoEntity;
 import jpa.jpa_shop.exception.NotSearchId;
 import jpa.jpa_shop.service.IFS.ItemServiceIFS;
-import jpa.jpa_shop.web.controller.dto.response.item.ItemListResponseDto;
+import jpa.jpa_shop.web.dto.response.item.ItemListResponseDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
