+++ src/main/java/jpa/jpa_shop/service/OrderService.java
@@ -11,8 +11,8 @@ import jpa.jpa_shop.domain.orders.Order;
 import jpa.jpa_shop.domain.orders.Repository.OrderRepository;
 import jpa.jpa_shop.exception.NoEntity;
 import jpa.jpa_shop.service.IFS.OrderServiceIFS;
-import jpa.jpa_shop.web.controller.dto.request.order.OrderSaveRequestDto;
-import jpa.jpa_shop.web.controller.dto.request.order.OrderSearchRequestDto;
+import jpa.jpa_shop.web.dto.request.order.OrderSaveRequestDto;
+import jpa.jpa_shop.web.dto.request.order.OrderSearchRequestDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
