+++ src/main/java/jpa/jpa_shop/web/dto/request/item/AlbumSaveRequestDto.java
@@ -1,14 +1,13 @@
-package jpa.jpa_shop.web.controller.dto.request.item;
+package jpa.jpa_shop.web.dto.request.item;
 
 import jpa.jpa_shop.domain.item.Album;
-import jpa.jpa_shop.domain.item.Movie;
 import lombok.Data;
 
 import javax.validation.constraints.Min;
 import javax.validation.constraints.NotEmpty;
 
 @Data
-public class MovieSaveRequestDto {
+public class AlbumSaveRequestDto {
     @NotEmpty(message = "이름을 입력해주세요.")
     private String name;
 
@@ -18,20 +17,21 @@ public class MovieSaveRequestDto {
     @Min(value = 1,message = "1개 이상이여야 합니다.")
     private int stockQuantity;
 
-    @NotEmpty(message = "감독을 입력해주세요.")
-    private String director;
+    @NotEmpty(message = "아티스트를 입력해주세요")
+    private String artist;
 
-    @NotEmpty(message = "주연 배우를 입력해주세요.")
-    private String actor;
+    @NotEmpty(message = "작곡가:xxx 작사가:xxx 형식으로 입력해주세요")
+    private String etc;
 
-    public Movie toEntity()
+    public Album toEntity()
     {
-        return Movie.builder()
+        return Album.builder()
                 .name(name)
                 .price(price)
                 .stockQuantity(stockQuantity)
-                .director(director)
-                .actor(actor)
+                .artist(artist)
+                .etc(etc)
                 .build();
     }
+
 }
