+++ src/main/java/jpa/jpa_shop/web/dto/response/item/AlbumUpdateResponseDto.java
@@ -1,14 +1,16 @@
-package jpa.jpa_shop.web.controller.dto.request.item;
+package jpa.jpa_shop.web.dto.response.item;
 
 import jpa.jpa_shop.domain.item.Album;
-import lombok.Data;
-import lombok.ToString;
+import lombok.*;
 
 import javax.validation.constraints.Min;
 import javax.validation.constraints.NotEmpty;
 
 @Data
-public class AlbumSaveRequestDto {
+public class AlbumUpdateResponseDto {
+
+    private Long id;
+
     @NotEmpty(message = "이름을 입력해주세요.")
     private String name;
 
@@ -24,15 +26,26 @@ public class AlbumSaveRequestDto {
     @NotEmpty(message = "작곡가:xxx 작사가:xxx 형식으로 입력해주세요")
     private String etc;
 
+    @Builder
+    public AlbumUpdateResponseDto(Long id, String name, int price, int stockQuantity, String artist, String etc) {
+        this.id = id;
+        this.name = name;
+        this.price = price;
+        this.stockQuantity = stockQuantity;
+        this.artist = artist;
+        this.etc = etc;
+    }
+
     public Album toEntity()
     {
-        return Album.builder()
+        Album album = Album.builder()
                 .name(name)
                 .price(price)
                 .stockQuantity(stockQuantity)
                 .artist(artist)
                 .etc(etc)
                 .build();
+        album.setId(id);
+        return album;
     }
-
 }
